requirements.txt
python-dotenv
Werkzeug~=3.1.5
Flask~=3.1.2
Flask-JWT-Extended~=4.7.1
PyMySQL~=1.1.2
gunicorn
flask-cors